#1343: implemented requested changes improved getStatusJson test method refactored PipUrlUpdaterTest
