+++ ide/src/test/java/com/devonfw/tools/ide/integrationtest/AbstractUrlUpdaterTest.java
@@ -1,30 +1,32 @@
 package com.devonfw.tools.ide.integrationtest;
 
-import java.io.BufferedReader;
-import java.nio.file.Files;
-import java.nio.file.Path;
-
 import org.assertj.core.api.Assertions;
 
-import com.devonfw.tools.ide.json.mapping.JsonMapping;
+import com.devonfw.tools.ide.url.model.file.UrlStatusFile;
 import com.devonfw.tools.ide.url.model.file.json.StatusJson;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.devonfw.tools.ide.url.model.folder.UrlEdition;
+import com.devonfw.tools.ide.url.model.folder.UrlRepository;
+import com.devonfw.tools.ide.url.model.folder.UrlTool;
+import com.devonfw.tools.ide.url.model.folder.UrlVersion;
 
 public class AbstractUrlUpdaterTest extends Assertions {
 
-  protected StatusJson getStatusJson(Path versionsPath) {
+  /**
+   * @param urlRepository {@link UrlRepository} to use
+   * @param toolName String of tool name
+   * @param editionName String of edition name
+   * @param versionName String of version name
+   * @return {@link StatusJson}
+   */
+  protected StatusJson retrieveStatusJson(UrlRepository urlRepository, String toolName, String editionName,
+      String versionName) {
 
-    ObjectMapper MAPPER = JsonMapping.create();
-    StatusJson statusJson = new StatusJson();
-    Path statusJsonPath = versionsPath.resolve("status.json");
-
-    if (Files.exists(statusJsonPath)) {
-      try (BufferedReader reader = Files.newBufferedReader(statusJsonPath)) {
-        statusJson = MAPPER.readValue(reader, StatusJson.class);
-      } catch (Exception e) {
-        throw new IllegalStateException("Failed to load " + statusJsonPath, e);
-      }
-    }
-    return statusJson;
+    UrlTool urlTool = new UrlTool(urlRepository, toolName);
+    UrlEdition urlEdition = new UrlEdition(urlTool, editionName);
+    UrlVersion urlVersion = new UrlVersion(urlEdition, versionName);
+    UrlStatusFile urlStatusFile = new UrlStatusFile(urlVersion);
+    urlStatusFile.load(false);
+    return urlStatusFile.getStatusJson();
   }
+
 }
